Create UserInputHandler.java.
Move the input logic from Controller into new class in directory in.
